should fix tests and refactor

+++ test/tornado_tests/test_motor_collection.py
@@ -264,6 +264,8 @@ class MotorCollectionTest(MotorTest):
 
     @gen_test
     async def test_async_create_encrypted_collection(self):
+        if pymongo.version_tuple < (4, 4, 0):
+            raise unittest.SkipTest("Requires PyMongo 4.4+")
         c = self.collection
         KMS_PROVIDERS = {"local": {"key": b"\x00" * 96}}
         self.cx.drop_database("db")
